+++ Assets/Model/Characters/Builders/HumanoidBuilder.cs
@@ -57,10 +57,10 @@ namespace Model.Characters
                 character.Type = c.Type;
                 this.BuildArmorHelper(character, c);
                 this.BuildWeaponHelper(character, c);
-                character.CurrentAP = character.GetCurrentStatValue(SecondaryStatsEnum.AP);
-                character.CurrentHP = character.GetCurrentStatValue(SecondaryStatsEnum.HP);
-                character.CurrentMorale = character.GetCurrentStatValue(SecondaryStatsEnum.Morale);
-                character.CurrentStamina = character.GetCurrentStatValue(SecondaryStatsEnum.Stamina);
+                character.CurrentAP = (int)character.GetCurrentStatValue(SecondaryStatsEnum.AP);
+                character.CurrentHP = (int)character.GetCurrentStatValue(SecondaryStatsEnum.HP);
+                character.CurrentMorale = (int)character.GetCurrentStatValue(SecondaryStatsEnum.Morale);
+                character.CurrentStamina = (int)character.GetCurrentStatValue(SecondaryStatsEnum.Stamina);
                 SpellMediator.Instance.SetCharacterSpells(character, c);
                 return character;
             }
