+++ Assets/View/GUI/HoverModal.cs
@@ -105,11 +105,11 @@ namespace View.GUI
             else
                 this._rWeaponSlider.maxValue = 0;
 
-            this._hpSlider.maxValue = c.GetCurrentStatValue(SecondaryStatsEnum.HP);
+            this._hpSlider.maxValue = (int)c.GetCurrentStatValue(SecondaryStatsEnum.HP);
             this._hpSlider.value = c.CurrentHP;
-            this._moraleSlider.maxValue = c.GetCurrentStatValue(SecondaryStatsEnum.Morale);
+            this._moraleSlider.maxValue = (int)c.GetCurrentStatValue(SecondaryStatsEnum.Morale);
             this._moraleSlider.value = c.CurrentMorale;
-            this._stamSlider.maxValue = c.GetCurrentStatValue(SecondaryStatsEnum.Stamina);
+            this._stamSlider.maxValue = (int)c.GetCurrentStatValue(SecondaryStatsEnum.Stamina);
             this._stamSlider.value = c.CurrentStamina;
 
         }
